% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIEF_convert_T.R
\name{BRIEF_convert_T}
\alias{BRIEF_convert_T}
\title{Converts BRIEF raw scores into T scores}
\usage{
BRIEF_convert_T(data)
}
\arguments{
\item{data}{dataframe (for Tim's project: a mild object), contains columns
with BRIEF raw scores for each of the 9 subscales and GEC}
}
\value{

}
\description{
Takes in data with BRIEF raw scores in it and converts them into BRIEF T-scores.
Make sure that subscales are named "BRIEF_(subscale)_Pre". Currently supports
the 9 clinical subscales (Inhibit, shift, working memory etc) and GEC. No
support yet for MI, BRI or other subscales. Designed for multiply-imputed data
}
\examples{
NULL
}
