% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fru.R
\name{Fru}
\alias{Fru}
\title{Estimate pooled F of a model from multiply imputed data}
\usage{
Fru(dataset, response, predictors, testpredictors)
}
\arguments{
\item{dataset}{mids object, a multiply-imputed dataset generated from mice}

\item{response}{character vector, specifying the name of the response variable in the dataset}

\item{predictors}{vector, specifying the name(s) of the predictors in the dataset}

\item{testpredictors}{vector, specifying the name(s) of predictors to be tested for significance}
}
\value{

}
\description{
Returns the F statistic of a regression model, given the response and
predictor variables. Not all variables need to be tested for significance
simultaneously, which allows F-testing of hierarchical regressions.
}
\examples{
mids <- mice::mice(mice::nhanes, m=5, maxit=1)
outcome <- "hyp"
predictors <- c("age", "bmi")
testpredictors <- c("age","bmi")
F1 <- Fru(mids, outcome, predictors, testpredictors)
}
