% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceout.R
\name{niceout}
\alias{niceout}
\title{Elegant tables for regressions on multiply imputed data}
\usage{
niceout(
  mids,
  filename,
  outcome,
  predictors1,
  predictors2,
  predictors3 = NULL,
  filepath = NULL
)
}
\arguments{
\item{mids}{mids object, a multiply-imputed dataset generated from mice}

\item{filename}{character vector, specify the name of the file (without extension)}

\item{outcome}{character vector, specifying the name of the response variable in the dataset}

\item{predictors1}{character vector, predictors in Step 1 of regression}

\item{predictors2}{character vector, predictors in Step 2 of regression}

\item{predictors3}{(OPTIONAL) character vector, predictors in Step 3 of regression}

\item{filepath}{string, specify full path to output directory in the format
"~/directory/subdirectory/"}
}
\value{

}
\description{
Outputs nicely-formatted regression tables in html format, containing standardised and
unstandardised regression coefficients with SEs, p values, t statistics,
F statistics, R-squared and R-squared change for regression models.
}
\examples{
mids <- mice::mice(mice::nhanes, m=5, maxit=1)

niceout(mids,"Hierarchical_regression_1",
outcome = "hyp",
predictors1 = c("age", "bmi"),
predictors2 = c("age", "bmi","chl"),
filepath = "D:/Desktop/")
}
