% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS_convert_T.R
\name{SRS_convert_T}
\alias{SRS_convert_T}
\title{Converts SRS raw scores into T scores}
\usage{
SRS_convert_T(data)
}
\arguments{
\item{data}{dataframe (for Tim's project: a mild object), contains columns
with BRIEF raw scores for each of the 9 subscales and GEC}
}
\value{

}
\description{
Takes in data with SRS raw scores in it and converts them into SRS T scores.
Currently works on data with the following subscales: Communication, Motivation,
Awareness, Cognition and SCI. Ensure that individual subscales are named
"SRS_RAW_Score_Pre_(subscale)" and the SCI is named "SRS_RAW_Total_Pre_SCI".
Designed for multiply imputed data.
}
\examples{
NULL
}
